requirements.txt
Authlib==1.3.2
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Deprecated==1.2.14
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
icalendar==5.0.13
idna==3.8
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mdurl==0.1.2
mutagen==1.47.0
mysql-connector-python==9.0.0
ordered-set==4.1.0
packaging==24.1
pycparser==2.22
pycryptodomex==3.20.0
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
rich==13.8.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.32
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.2
websockets==13.0.1
Werkzeug==3.0.4
wrapt==1.16.0
yt-dlp==2024.8.6
PyMySQL==1.0.3
qrcode==7.4.2